+++ libqpdf/QPDF_Array.cc
@@ -49,13 +49,13 @@ QPDF_Array::getTypeName() const
 int
 QPDF_Array::getNItems() const
 {
-    return (int)this->items.size();
+    return this->items.size();
 }
 
 QPDFObjectHandle
 QPDF_Array::getItem(int n) const
 {
-    if ((n < 0) || (n >= (int)this->items.size()))
+    if ((n < 0) || (n >= static_cast<int>(this->items.size())))
     {
 	throw std::logic_error(
 	    "INTERNAL ERROR: bounds error accessing QPDF_Array element");
@@ -87,7 +87,7 @@ void
 QPDF_Array::insertItem(int at, QPDFObjectHandle const& item)
 {
     // As special case, also allow insert beyond the end
-    if ((at < 0) || (at > (int)this->items.size()))
+    if ((at < 0) || (at > static_cast<int>(this->items.size())))
     {
 	throw std::logic_error(
 	    "INTERNAL ERROR: bounds error accessing QPDF_Array element");
